==> 03-Data-Preparation.ipynb <==
# Data Preparation and Market Data

### Text Mining on Earnings Calls during a Pandemic as a Means to Predict End-Of-The-Month Stock Performances
####  Olin School of Business <br> Jose Luis Rodriguez  <br> jlr@wustl.edu <br> Fall 2021

import intrinio_sdk
import configparser as cp
import pandas as pd


**Intrinio API: Secure method to load API credentials**

cfg = cp.ConfigParser()
cfg.read('resources/credentials.cfg')

API_KEY = cfg['intrinio']['app_key']

intrinio_sdk.ApiClient().configuration.api_key['api_key'] = API_KEY

security_api = intrinio_sdk.SecurityApi()

## Hotels, Restaurants and Leasure Stocks Tickers

sdf = pd.read_csv('data/hrl_mrk20.csv')

tickers = sdf['related'].apply(lambda x:x.split(':')[1] if len(x.split(':'))>1 else '')
sdf['ticker'] = tickers
tickers = tickers.reset_index()
tickers.columns = ['index','ticker']
tickers = tickers.drop(columns = ['index'])

# date | Return prices on or after the date (optional)
start_date = '2021-01-01'

# date | Return prices on or before the date (optional)
end_date = '2021-12-31'

# str | Return stock prices in the given frequency (optional) (default to daily)
frequency = 'monthly'

dfs = []
missing = []
for ticker in tickers['ticker'].unique():
    try:
        next_page = ''
        response = security_api.get_security_stock_prices(ticker,
                                                          start_date = start_date,
                                                          end_date = end_date,
                                                          frequency = frequency)
        df = [p.to_dict() for p in response.stock_prices]
        next_page = response.next_page
        if next_page != None:
            response = security_api.get_security_stock_prices(ticker,
                                                              start_date = start_date,
                                                              end_date = end_date,
                                                              frequency = frequency,
                                                              next_page = next_page)    
            df.extend(p.to_dict() for p in response.stock_prices)
        df = pd.DataFrame.from_dict(df)
        df['secid'] = ticker
        dfs.append(df)
    except:
        missing.append(ticker)

sdf = pd.concat(dfs)
sdf.index = pd.DatetimeIndex(sdf['date'])
sdf = sdf.drop('date', axis=1)
sdf.index.name = None

#SORT DATETIME INDEX
sdf = sdf.sort_index()

sdf.head()

sdf.to_csv('data/trs_mnt-2021.csv')

## Market + Earnings Data Merge

hrl20 = pd.read_csv('data/source/hrl_2020.csv')
related = hrl20['related'].apply(lambda x: x.split(':')[1] if len(x.split(':')) > 1 else x.split(':')[0])
hrl20.loc[:,'related'] = related
hrl20.loc[:,'date'] = pd.to_datetime(hrl20['date'])
hrl20['month'] = hrl20['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

hrl_mnt20 = pd.read_csv('data/source/hrl_mnt-2020.csv')
hrl_mnt20 = hrl_mnt20.rename(columns = {'Unnamed: 0':'date'})
htl_mnt20 = hrl_mnt20.drop(columns = ['intraperiod', 'frequency', 'fifty_two_week_high',
                                      'fifty_two_week_low', 'factor', 'split_ratio', 'dividend'])
htl_mnt20.loc[:,'date'] = pd.to_datetime(htl_mnt20['date'])
htl_mnt20['month'] = htl_mnt20['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

hrl21 = pd.read_csv('data/source/hrl_2021.csv')
related = hrl21['related'].apply(lambda x: x.split(':')[1] if len(x.split(':')) > 1 else x.split(':')[0])
hrl21.loc[:,'related'] = related
hrl21.loc[:,'date'] = pd.to_datetime(hrl21['date'])
hrl21['month'] = hrl21['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

hrl_mnt21 = pd.read_csv('data/source/hrl_mnt-2021.csv')
hrl_mnt21 = hrl_mnt21.rename(columns = {'Unnamed: 0':'date'})
htl_mnt21 = hrl_mnt21.drop(columns = ['intraperiod', 'frequency', 'fifty_two_week_high',
                                      'fifty_two_week_low', 'factor', 'split_ratio', 'dividend'])
htl_mnt21.loc[:,'date'] = pd.to_datetime(htl_mnt21['date'])
htl_mnt21['month'] = htl_mnt21['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

bind = []
for sec in hrl20['related'].unique():
    dfco = hrl20[hrl20['related'] == sec]
    dfmnt = htl_mnt20[htl_mnt20['secid'] == sec]
    if dfmnt.shape[0] > 1:
        dfco = dfco.merge(dfmnt, on = 'month')
        dfco = dfco.drop(columns = ['secid', 'month', 'doc_type', 'link'])
        dfco = dfco.rename(columns = {'date_x':'date_earnings', 'date_y':'date_market'})
        dfco = dfco[['date_earnings', 'date_market', 'time', 'related', 'company',
                     'speakers_info', 'speakers_transcript',
                     'speakers_number', 'corpus', 'adj_open', 'adj_high', 'adj_low',
                     'adj_close', 'adj_volume', 'change', 'percent_change']]
        dfco = dfco.sort_values(by = 'date_earnings')
        dfco['direction'] = dfco['percent_change'].apply(lambda x: 'positive' if x>0 else 'negative')        
        bind.append(dfco)
hrl20_df = pd.concat(bind)
hrl20_df.shape

bind = []
for sec in hrl21['related'].unique():
    dfco = hrl21[hrl21['related'] == sec]
    dfmnt = htl_mnt21[htl_mnt21['secid'] == sec]
    if dfmnt.shape[0] > 1:
        dfco = dfco.merge(dfmnt, on = 'month')
        dfco = dfco.drop(columns = ['secid', 'month', 'doc_type', 'link'])
        dfco = dfco.rename(columns = {'date_x':'date_earnings', 'date_y':'date_market'})
        dfco = dfco[['date_earnings', 'date_market', 'time', 'related', 'company',
                     'speakers_info', 'speakers_transcript',
                     'speakers_number', 'corpus', 'adj_open', 'adj_high', 'adj_low',
                     'adj_close', 'adj_volume', 'change', 'percent_change']]
        dfco = dfco.sort_values(by = 'date_earnings')
        dfco['direction'] = dfco['percent_change'].apply(lambda x: 'positive' if x>0 else 'negative')        
        bind.append(dfco)
hrl21_df = pd.concat(bind)
hrl21_df.shape

trs20 = pd.read_csv('data/source/trs_2020.csv')
related = trs20['related'].apply(lambda x: x.split(':')[1] if len(x.split(':')) > 1 else x.split(':')[0])
trs20.loc[:,'related'] = related
trs20.loc[:,'date'] = pd.to_datetime(trs20['date'])
trs20['month'] = trs20['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

trs_mnt20 = pd.read_csv('data/source/trs_mnt-2020.csv')
trs_mnt20 = trs_mnt20.rename(columns = {'Unnamed: 0':'date'})
trs_mnt20 = trs_mnt20.drop(columns = ['intraperiod', 'frequency', 'fifty_two_week_high',
                                      'fifty_two_week_low', 'factor', 'split_ratio', 'dividend'])
trs_mnt20.loc[:,'date'] = pd.to_datetime(trs_mnt20['date'])
trs_mnt20['month'] = trs_mnt20['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

trs21 = pd.read_csv('data/source/trs_2021.csv')
related = trs21['related'].apply(lambda x: x.split(':')[1] if len(x.split(':')) > 1 else x.split(':')[0])
trs21.loc[:,'related'] = related
trs21.loc[:,'date'] = pd.to_datetime(trs21['date'])
trs21['month'] = trs21['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

trs_mnt21 = pd.read_csv('data/source/trs_mnt-2021.csv')
trs_mnt21 = trs_mnt21.rename(columns = {'Unnamed: 0':'date'})
trs_mnt21 = trs_mnt21.drop(columns = ['intraperiod', 'frequency', 'fifty_two_week_high',
                                      'fifty_two_week_low', 'factor', 'split_ratio', 'dividend'])
trs_mnt21.loc[:,'date'] = pd.to_datetime(trs_mnt21['date'])
trs_mnt21['month'] = trs_mnt21['date'].apply(lambda x:str(x.month).zfill(2) + '-' + '2020' )

bind = []
for sec in trs20['related'].unique():
    dfco = trs20[trs20['related'] == sec]
    dfmnt = trs_mnt20[trs_mnt20['secid'] == sec]
    if dfmnt.shape[0] > 1:
        dfco = dfco.merge(dfmnt, on = 'month')
        dfco = dfco.drop(columns = ['secid', 'month', 'doc_type', 'link'])
        dfco = dfco.rename(columns = {'date_x':'date_earnings', 'date_y':'date_market'})
        dfco = dfco[['date_earnings', 'date_market', 'time', 'related', 'company',
                     'speakers_info', 'speakers_transcript',
                     'speakers_number', 'corpus', 'adj_open', 'adj_high', 'adj_low',
                     'adj_close', 'adj_volume', 'change', 'percent_change']]
        dfco = dfco.sort_values(by = 'date_earnings')
        dfco['direction'] = dfco['percent_change'].apply(lambda x: 'positive' if x>0 else 'negative')        
        bind.append(dfco)
trs20_df = pd.concat(bind)
trs20_df.shape

bind = []
for sec in trs21['related'].unique():
    dfco = trs21[trs21['related'] == sec]
    dfmnt = trs_mnt21[trs_mnt21['secid'] == sec]
    if dfmnt.shape[0] > 1:
        dfco = dfco.merge(dfmnt, on = 'month')
        dfco = dfco.drop(columns = ['secid', 'month', 'doc_type', 'link'])
        dfco = dfco.rename(columns = {'date_x':'date_earnings', 'date_y':'date_market'})
        dfco = dfco[['date_earnings', 'date_market', 'time', 'related', 'company',
                     'speakers_info', 'speakers_transcript',
                     'speakers_number', 'corpus', 'adj_open', 'adj_high', 'adj_low',
                     'adj_close', 'adj_volume', 'change', 'percent_change']]
        dfco = dfco.sort_values(by = 'date_earnings')
        dfco['direction'] = dfco['percent_change'].apply(lambda x: 'positive' if x>0 else 'negative')        
        bind.append(dfco)
trs21_df = pd.concat(bind)
trs21_df.shape

hrl20_df.to_csv('data/hrl_mrk20.csv', index = False, doublequote = True)
hrl21_df.to_csv('data/hrl_mrk21.csv', index = False, doublequote = True)
trs20_df.to_csv('data/trs_mrk20.csv', index = False, doublequote = True)
trs21_df.to_csv('data/trs_mrk21.csv', index = False, doublequote = True)
