==> 05-Earnings-Themes.ipynb <==
# Earnings Calls - Topic Modeling

### Text Mining on Earnings Calls during a Pandemic as a Means to Predict End-Of-The-Month Stock Performances
####  Olin School of Business <br> Jose Luis Rodriguez  <br> jlr@wustl.edu <br> Fall 2021

import sys
import nltk
import numpy as np
import pandas as pd
from sklearn import metrics
from sklearn.feature_extraction.text import  CountVectorizer #bag-of-words vectorizer 
from sklearn.decomposition import LatentDirichletAllocation #package for LDA

# Plotting tools
import pyLDAvis
import pyLDAvis.sklearn

import matplotlib.pyplot as plt

%%capture
#define text normalization function
%run ./'02-Normalization.ipynb' #defining text normalization function

def display_topics(model, feature_names, no_top_words):
    for topic_idx, topic in enumerate(model.components_):
        print("Topic %d:" % (topic_idx))
        print(" ".join([feature_names[i]
                        for i in topic.argsort()[:-no_top_words - 1:-1]]))
        
def get_topic_words(vectorizer, lda_model, n_words):
    keywords = np.array(vectorizer.get_feature_names())
    topic_words = []
    for topic_weights in lda_model.components_:
        top_word_locs = (-topic_weights).argsort()[:n_words]
        topic_words.append(keywords.take(top_word_locs).tolist())
    return topic_words

### Hotels, Restaurants and Leisure

def earnings_lda(segment, max_features = 1000, components = 4, topic_prior = 0.1, word_prior = 0.1):
    bow_vectorizer = CountVectorizer(max_features = max_features)
    for secid in segment['related'].unique():
        data = segment[segment['related'] == secid]
        corpus = []
        for n in range(data.shape[0]):
            cdata = data.iloc[n].to_dict()
            corpus.extend([i.strip() for i in cdata.split('\n') if i.strip() != ''])
        normalized_corpus = normalize_corpus(corpus)
        bow_corpus = bow_vectorizer.fit_transform(normalized_corpus)
        lda_corpus = LatentDirichletAllocation(n_components = components,
                                               max_iter = 100,
                                               doc_topic_prior = topic_prior,
                                               topic_word_prior = word_prior).fit(bow_corpus)
        lda_topic_weights = lda_corpus.transform(bow_corpus)
        doc_names = ["Doc_" + str(i) for i in range(len(normalized_corpus))]
        topic_names = ["Topic_" + str(i) for i in range(components)]
        topic_df = pd.DataFrame(np.round(lda_topic_weights, components),
                                 columns = topic_names,
                                 index = doc_names)
        topic_df['secid'] = secid
        topic_df['direction'] = cdata['direction']
        topic_df['date_market'] = cdata['date_market']
        topic_df['speakers_number'] = cdata['speakers_number']
        topic_df['percent_change'] = cdata['percent_change']
        return topic_df

hrl_df = pd.read_csv('data/hrl_mrk21.csv')
hrl_df['date_market'] = pd.to_datetime(hrl_df['date_market'])
hrl_df['market_month'] = hrl_df['date_market'].apply(lambda i:str(i.month).zfill(2) + '-2021')

hrl_company_topics = earnings_lda(hrl_df)

## Monthly Earnings Calls Topics

netd = pd.DataFrame({'count':hrl_df.groupby(['market_month','direction'])['direction'].count()}).reset_index()
neg = netd[netd['direction'] == 'negative'].reset_index(drop=True)
direction = netd[netd['direction'] == 'positive'].reset_index(drop=True)
direction['net_change'] = (direction['count'] - neg['count']).reset_index(drop=True)
direction['market'] = direction['net_change'].apply(lambda i:'negative' if  i< 0 else 'positive')   
direction = direction[['market_month','net_change','market']]

month_lda = []
components = 4
for month in hrl_df['market_month'].unique():
    mdata = hrl_df[hrl_df['market_month'] == month].copy(deep=True)
    market = direction[direction['market_month'] == month]['market']
    net_change = direction[direction['market_month'] == month]['net_change']
    cps = []
    for n in range(mdata.shape[0]):
        cdata = mdata.iloc[n].to_dict()
        corpus = cdata['corpus']
        cps.extend([i.strip() for i in corpus.split('\n') if i.strip() != ''])
    normalized_corpus = normalize_corpus(cps)
    bow_vectorizer = CountVectorizer(max_features=2000)
    bow_corpus = bow_vectorizer.fit_transform(normalized_corpus)
    lda_corpus = LatentDirichletAllocation(n_components = components,
                                           max_iter = 100,
                                           doc_topic_prior = 0.2,
                                           topic_word_prior = 0.2).fit(bow_corpus)
    word_weights = lda_corpus.components_ / lda_corpus.components_.sum(axis=1)[:, np.newaxis]
    word_weights_df = pd.DataFrame(word_weights.T, 
                                   index = bow_vectorizer.get_feature_names(), 
                                   columns = ["Topic_" + str(i) for i in range(components)])
    word_weights_df = word_weights_df.sort_values(by = 'Topic_0', ascending = False)
    word_weights_df.loc[:,'signal'] = market.to_string(header=False, index=False)
    word_weights_df.loc[:,'month'] = month
    word_weights_df.loc[:,'net_change'] = net_change.to_string(header=False, index=False)    
    month_lda.append(word_weights_df)

words_df = pd.concat(month_words)
words_df = words_df.reset_index()

words_df[words_df['index'] == 'inflation']

words_df.to_csv('data/words_topics.csv', index=False)

corpus = []
for n in range(hrl_df.shape[0]):
    cps = hrl_df.iloc[n]['corpus']
    cps = [i.strip() for i in cps.split('\n') if i.strip() != '']
    corpus.extend(cps)

normalized_corpus = normalize_corpus(corpus)

#define the bag-of-words vectorizer:
bow_vectorizer = CountVectorizer(max_features=1000)
#vectorize the normalized data:
bow_corpus = bow_vectorizer.fit_transform(normalized_corpus)

components = 4
lda_corpus = LatentDirichletAllocation(n_components = components,
                                       max_iter = 100,
                                       doc_topic_prior = 0.2,
                                       topic_word_prior = 0.2).fit(bow_corpus)

no_top_words = 15
display_topics(lda_corpus, bow_vectorizer.get_feature_names(), no_top_words)

word_weights = lda_corpus.components_ / lda_corpus.components_.sum(axis=1)[:, np.newaxis]
word_weights_df = pd.DataFrame(word_weights.T, 
                               index = bow_vectorizer.get_feature_names(), 
                               columns = ["Topic_" + str(i) for i in range(components)])
word_weights_df = word_weights_df.sort_values(by = 'Topic_0', ascending = False)

### Topic Model Visualization


#prepare to display result in the Jupyter notebook
pyLDAvis.enable_notebook()

# mds is a function to use for visualizing the "distance" between topics]
vis_data = pyLDAvis.sklearn.prepare(lda_corpus,
                                     bow_corpus,
                                     bow_vectorizer,
                                     mds = 'tsne')

pyLDAvis.save_html(vis_data, 'lda-2021.html')
